Add separate file for `CertifyKey` command.
