+++ dpe/src/dpe_instance.rs
@@ -7,15 +7,11 @@ Abstract:
 use crate::{
     _set_flag,
     commands::{
-        CertifyKeyCmd, Command, CommandExecution, DestroyCtxCmd, ExtendTciCmd, InitCtxCmd,
-        RotateCtxCmd, TagTciCmd,
+        Command, CommandExecution, DestroyCtxCmd, ExtendTciCmd, InitCtxCmd, RotateCtxCmd, TagTciCmd,
     },
-    response::{CertifyKeyResp, DpeErrorCode, GetProfileResp, NewHandleResp, Response},
-    x509::{EcdsaPub, EcdsaSignature, MeasurementData, Name, X509CertWriter},
-    DPE_PROFILE, HANDLE_SIZE, MAX_CERT_SIZE, MAX_HANDLES,
+    response::{DpeErrorCode, GetProfileResp, NewHandleResp, Response},
+    DPE_PROFILE, HANDLE_SIZE, MAX_HANDLES,
 };
-use core::convert::TryFrom;
-use core::fmt::{Error, Write};
 use core::marker::PhantomData;
 use core::mem::size_of;
 use crypto::{Crypto, Hasher};
@@ -197,123 +193,6 @@ impl<C: Crypto> DpeInstance<'_, C> {
         }))
     }
 
-    pub fn certify_key(
-        &mut self,
-        locality: u32,
-        cmd: &CertifyKeyCmd,
-    ) -> Result<CertifyKeyResp, DpeErrorCode> {
-        let idx = self
-            .get_active_context_pos(&cmd.handle, locality)
-            .ok_or(DpeErrorCode::InvalidHandle)?;
-
-        // Make sure the command is coming from the right locality.
-        if self.contexts[idx].locality != locality {
-            return Err(DpeErrorCode::InvalidHandle);
-        }
-
-        // Get TCI Nodes
-        const INITIALIZER: TciNodeData = TciNodeData::new();
-        let mut nodes = [INITIALIZER; MAX_HANDLES];
-        let tcb_count = self.get_tcb_nodes(&self.contexts[idx], &mut nodes)?;
-
-        // Hash TCI Nodes
-        let mut tci_bytes = [0u8; MAX_HANDLES * size_of::<TciNodeData>()];
-        let mut tci_offset = 0;
-        for n in &nodes[..tcb_count] {
-            tci_offset += n.serialize(&mut tci_bytes[tci_offset..])?;
-        }
-
-        let mut digest = [0; DPE_PROFILE.get_hash_size()];
-        C::hash(DPE_PROFILE.alg_len(), &tci_bytes[..tci_offset], &mut digest)
-            .map_err(|_| DpeErrorCode::InternalError)?;
-
-        // Derive CDI and public key
-        let cdi = C::derive_cdi(DPE_PROFILE.alg_len(), &digest, b"DPE")
-            .map_err(|_| DpeErrorCode::InternalError)?;
-        let mut pub_key = EcdsaPub::default();
-        C::derive_ecdsa_pub(
-            DPE_PROFILE.alg_len(),
-            &cdi,
-            &cmd.label,
-            b"ECC",
-            &mut pub_key.x,
-            &mut pub_key.y,
-        )
-        .map_err(|_| DpeErrorCode::InternalError)?;
-
-        // Hash the public key for the serialNumber name field
-        let mut pub_bytes = [0u8; size_of::<EcdsaPub>()];
-        let pub_offset = pub_key.serialize(&mut pub_bytes)?;
-        let mut pub_digest = [0u8; DPE_PROFILE.get_hash_size()];
-        C::hash(
-            DPE_PROFILE.alg_len(),
-            &pub_bytes[..pub_offset],
-            &mut pub_digest,
-        )
-        .map_err(|_| DpeErrorCode::InternalError)?;
-
-        // TODO: Let the platform specify issuer name
-        let issuer_name = Name {
-            cn: b"DPE Issuer",
-            serial: b"000000",
-        };
-
-        let mut subject_sn_str = [0u8; DPE_PROFILE.get_hash_size() * 2];
-        let mut w = BufWriter {
-            buf: &mut subject_sn_str,
-            offset: 0,
-        };
-        w.write_hex_str(&pub_digest)?;
-
-        let subject_name = Name {
-            cn: b"DPE Leaf",
-            serial: &subject_sn_str,
-        };
-
-        let measurements = MeasurementData {
-            _label: &cmd.label,
-            tci_nodes: &nodes[..tcb_count],
-        };
-
-        // Get certificate
-        let mut tbs_buffer = [0u8; MAX_CERT_SIZE];
-        let mut tbs_writer = X509CertWriter::new(&mut tbs_buffer);
-        let mut bytes_written = tbs_writer.encode_ecdsa_tbs(
-            /*serial=*/ &pub_digest,
-            &issuer_name,
-            &subject_name,
-            &pub_key,
-            &measurements,
-        )?;
-
-        let mut tbs_digest = [0u8; DPE_PROFILE.get_hash_size()];
-        C::hash(
-            DPE_PROFILE.alg_len(),
-            &tbs_buffer[..bytes_written],
-            &mut tbs_digest,
-        )
-        .map_err(|_| DpeErrorCode::InternalError)?;
-        let mut sig = EcdsaSignature::default();
-        C::ecdsa_sign_with_alias(DPE_PROFILE.alg_len(), &tbs_digest, &mut sig.r, &mut sig.s)
-            .map_err(|_| DpeErrorCode::InternalError)?;
-
-        let mut cert = [0u8; MAX_CERT_SIZE];
-        let mut cert_writer = X509CertWriter::new(&mut cert);
-        bytes_written = cert_writer.encode_ecdsa_certificate(&tbs_buffer[..bytes_written], &sig)?;
-        let cert_size = u32::try_from(bytes_written).map_err(|_| DpeErrorCode::InternalError)?;
-
-        // Rotate handle if it isn't the default
-        self.roll_onetime_use_handle(idx)?;
-
-        Ok(CertifyKeyResp {
-            new_context_handle: self.contexts[idx].handle,
-            derived_pubkey_x: pub_key.x,
-            derived_pubkey_y: pub_key.y,
-            cert_size,
-            cert,
-        })
-    }
-
     /// Deserializes the command and executes it.
     ///
     /// # Arguments
@@ -334,9 +213,7 @@ impl<C: Crypto> DpeInstance<'_, C> {
             Command::GetProfile => Ok(Response::GetProfile(self.get_profile()?)),
             Command::InitCtx(cmd) => cmd.execute(self, locality),
             Command::DeriveChild(cmd) => cmd.execute(self, locality),
-            Command::CertifyKey(context) => {
-                Ok(Response::CertifyKey(self.certify_key(locality, &context)?))
-            }
+            Command::CertifyKey(cmd) => cmd.execute(self, locality),
             Command::RotateCtx(cmd) => self.rotate_context(locality, &cmd),
             Command::DestroyCtx(context) => self.destroy_context(locality, &context),
             Command::ExtendTci(cmd) => self.extend_tci(locality, &cmd),
@@ -440,34 +317,6 @@ impl<C: Crypto> DpeInstance<'_, C> {
     }
 }
 
-struct BufWriter<'a> {
-    buf: &'a mut [u8],
-    offset: usize,
-}
-
-impl Write for BufWriter<'_> {
-    fn write_str(&mut self, s: &str) -> Result<(), Error> {
-        if s.len() > self.buf.len().saturating_sub(self.offset) {
-            return Err(Error::default());
-        }
-
-        self.buf[self.offset..self.offset + s.len()].copy_from_slice(s.as_bytes());
-        self.offset += s.len();
-
-        Ok(())
-    }
-}
-
-impl BufWriter<'_> {
-    fn write_hex_str(&mut self, src: &[u8]) -> Result<(), DpeErrorCode> {
-        for &b in src {
-            write!(self, "{b:02x}").map_err(|_| DpeErrorCode::InternalError)?;
-        }
-
-        Ok(())
-    }
-}
-
 #[repr(transparent)]
 #[derive(Copy, Clone)]
 pub struct TciMeasurement(pub [u8; DPE_PROFILE.get_tci_size()]);
@@ -567,7 +416,7 @@ impl TciNodeData {
         }
     }
 
-    fn serialize(&self, dst: &mut [u8]) -> Result<usize, DpeErrorCode> {
+    pub fn serialize(&self, dst: &mut [u8]) -> Result<usize, DpeErrorCode> {
         if dst.len() < size_of::<Self>() {
             return Err(DpeErrorCode::InternalError);
         }
@@ -705,9 +554,6 @@ pub mod tests {
     use crate::DpeProfile;
     use crate::{commands::CommandHdr, CURRENT_PROFILE_VERSION};
     use crypto::OpensslCrypto;
-    use x509_parser::nom::Parser;
-    use x509_parser::prelude::X509CertificateParser;
-    use x509_parser::prelude::*;
     use zerocopy::AsBytes;
 
     const SUPPORT: Support = Support {
@@ -911,36 +757,6 @@ pub mod tests {
                 }
             )
         );
-    }
-
-    #[test]
-    fn test_certify_key() {
-        let mut dpe =
-            DpeInstance::<OpensslCrypto>::new(Support::default(), &TEST_LOCALITIES).unwrap();
-
-        let init_resp = match InitCtxCmd::new_use_default()
-            .execute(&mut dpe, TEST_LOCALITIES[0])
-            .unwrap()
-        {
-            Response::InitCtx(resp) => resp,
-            _ => panic!("Incorrect return type."),
-        };
-        let certify_cmd = CertifyKeyCmd {
-            handle: init_resp.handle,
-            flags: 0,
-            label: [0; DPE_PROFILE.get_hash_size()],
-        };
-
-        let certify_resp = dpe.certify_key(TEST_LOCALITIES[0], &certify_cmd).unwrap();
-        assert_ne!(certify_resp.cert_size, 0);
-
-        let mut parser = X509CertificateParser::new().with_deep_parse_extensions(false);
-        match parser.parse(&certify_resp.cert[..certify_resp.cert_size.try_into().unwrap()]) {
-            Ok((_, cert)) => {
-                assert_eq!(cert.version(), X509Version::V3);
-            }
-            Err(e) => panic!("x509 parsing failed: {:?}", e),
-        };
     }
 
     #[test]
